Rename "property" to "attribute"
We inconsistently referred to "a quality associated with nodes or edges"
(e.g., atomic number) as "property" and "attribute".
I renamed all occurrences of "property" to "attribute",
since it matches the NetworkX naming (e.g., `nx.set_node_attributes()`)
and to avoid collision with the built-in `property()` function in Python
(https://docs.python.org/3/library/functions.html#built-in-functions).